+++ changelog.md
@@ -11,7 +11,11 @@
   * Reduced items enchantabilities by ~25%
   * Enchantments
     * Cost per level has been reduced by ~1 (**you should regenerate the enchantment_data.json file**)
-    * Enduring is now III levels
+    * Enduring is now III levels (was II)
+    * Veining is now III levels (was II)
+    * Haste is now V levels (was II) with 0.8 tool efficiency per level
+    * Part breaker is now 1 level only, with 10% chance
+    * Armor piercer now reduces overall attack damage instead of the tool's
 * Heavily increased durium drops (2\~4 -> 4\~6)
 
 ### Fixes